+++ cedar-policy-symcc/src/lib.rs
@@ -191,6 +191,11 @@ impl CompiledPolicy {
         self.policy.effect()
     }
 
+    /// Get the (post-typecheck) `Policy` that this `CompiledPolicy` represents
+    pub fn policy(&self) -> &cedar_policy_core::ast::Policy {
+        &self.policy.policy
+    }
+
     /// Convert a `CompiledPolicy` to a `CompiledPolicies` representing a
     /// singleton policyset with just that policy.
     ///
@@ -251,6 +256,11 @@ impl CompiledPolicies {
                 symenv,
             )?,
         })
+    }
+
+    /// Get the (post-typecheck) `PolicySet` that this `CompiledPolicies` represents
+    pub fn policies(&self) -> &cedar_policy_core::ast::PolicySet {
+        &self.policies.policies
     }
 }